Move resources to authorization pkg, cleanup unneeded old files (#201)
