--- .travis.yml
@@ -1,18 +1,0 @@
-language: go
-
-matrix:
-  include:
-  - go: 1.13.x
-  - go: 1.12.x
-    env: GO111MODULE=on
-    
-install:
-- go get github.com/mattn/goveralls
-- go get -u github.com/client9/misspell/cmd/misspell
-script:
-- make test
-- if ([[ ${TRAVIS_BRANCH} == "master" ]] && [[ ${TRAVIS_EVENT_TYPE} == "push" ]]); then
-    go get github.com/mattn/goveralls;
-    goveralls -service=travis-ci;
-    make bench;
-  fi
