+++ doc.go
@@ -70,22 +70,6 @@ var (
 	)
 )
 
-// Resource represents a url resource to protect
-type Resource struct {
-	// URL the url for the resource
-	URL string `json:"uri" yaml:"uri"`
-	// Methods the method type
-	Methods []string `json:"methods" yaml:"methods"`
-	// WhiteListed permits the prefix through
-	WhiteListed bool `json:"white-listed" yaml:"white-listed"`
-	// RequireAnyRole indicates that ANY of the roles are required, the default is all
-	RequireAnyRole bool `json:"require-any-role" yaml:"require-any-role"`
-	// Roles the roles required to access this url
-	Roles []string `json:"roles" yaml:"roles"`
-	// Groups is a list of groups the user is in
-	Groups []string `json:"groups" yaml:"groups"`
-}
-
 // Config is the configuration for the proxy
 type Config struct {
 	// ConfigFile is the binding interface
@@ -125,7 +109,7 @@ type Config struct {
 	// UpstreamCA is the path to a CA certificate in PEM format to validate the upstream certificate
 	UpstreamCA string `json:"upstream-ca" yaml:"upstream-ca" usage:"the path to a file container a CA certificate to validate the upstream tls endpoint" env:"UPSTREAM_CA"`
 	// Resources is a list of protected resources
-	Resources []*Resource `json:"resources" yaml:"resources" usage:"list of resources 'uri=/admin*|methods=GET,PUT|roles=role1,role2'"`
+	Resources []*authorization.Resource `json:"resources" yaml:"resources" usage:"list of resources 'uri=/admin*|methods=GET,PUT|roles=role1,role2'"`
 	// Headers permits adding customs headers across the board
 	Headers map[string]string `json:"headers" yaml:"headers" usage:"custom headers to the upstream request, key=value"`
 	// PreserveHost preserves the host header of the proxied request in the upstream request
